Added a Data type setting
I spent most of the day trying to learn about LBFGS and quasi Newton
optimization methods. However, I decided to take a break from that to
add a new feature to the code. In particular, I added a setting which
controls the data type of all tensors in my code.

I had hard coded everything to be a single precision float before.
However, I was quickly approaching machine precision (I think). I,
therefore, decided to add a setting to control precision (just to see
if it'd help). This change forced me to modify a lot of the code (which
had been hard coded to work with float32's).

Anyway, the change worked.

+++ Code/Extraction.py
@@ -263,7 +263,7 @@ def Generate_Library(
     # below).
     num_rows : int = Coords.shape[0];
     num_cols : int = num_multi_indices.sum();
-    Library : np.array = np.ones((num_rows, num_cols), dtype = np.float32);
+    Library : np.array = np.ones((num_rows, num_cols), dtype = np.float64);
 
     # Evaluate u, du/dx,... at each point.
     (du_dt, diu_dxi) = Evaluate_u_Derivatives(
@@ -294,7 +294,7 @@ def Generate_Library(
             # Cycle through the sub-indices of this multi-index.
             for j in range(degree):
                 Library[:, position] = (Library[:, position]*
-                                        diu_dxi[:, multi_indices[i, j]].detach().squeeze().numpy());
+                                        diu_dxi[:, multi_indices[i, j]].detach().squeeze().numpy().astype(dtype = np.float64));
 
             # Increment position
             position += 1;
@@ -304,7 +304,7 @@ def Generate_Library(
 
     # All done, the library is now populated! Package everything together and
     # return.
-    return (N_NN_batch.detach().squeeze().numpy(),
+    return (N_NN_batch.detach().squeeze().numpy().astype(dtype = np.float64),
             Library,
             num_multi_indices,
             multi_indices_list);
